update get_neuron ages to use uint32 and to count in batches

+++ src/experiments/PS_factory.py
@@ -137,7 +137,7 @@ def run_experiment_PS_factory(alg, alg_params, text, B, T, N, epochs, tasks, see
 
     # initialize neuron_ages_array
     if save_neuron_ages:
-        neuron_ages_array = np.zeros((3, int(train_steps_per_task * tasks), int(config.n_embd * 4)))
+        neuron_ages_array = np.zeros((3, int(train_steps_per_task * tasks), int(config.n_embd * 4)), dtype = np.uint32)
 
     # Iniitialize loss_array
     loss_array = np.zeros(train_steps_per_task * tasks)
